# feat(SMT): reverse mutations generation, mutations serialization