--- a/Makefile
+++ b/Makefile
@@ -81,6 +81,10 @@ build-sve: ## Build with sve support
 
 # --- benchmarking --------------------------------------------------------------------------------
 
-.PHONY: bench-tx
-bench-tx: ## Run crypto benchmarks
-	cargo bench
+.PHONY: bench
+bench: ## Run crypto benchmarks
+	cargo bench --features="concurrent"
+
+.PHONY: bench-smt-concurrent
+bench-smt-concurrent: ## Run SMT benchmarks with concurrent feature
+	cargo run --release --features executable -- --size 1000000