--- a/src/main.rs
+++ b/src/main.rs
@@ -75,7 +75,7 @@ pub fn insertion(tree: &mut Smt) -> Result<(), MerkleError> {
     }
 
     println!(
-        "An average insertion time measured by {NUM_INSERTIONS} inserts into an SMT with {size} leaves is {:.1} μs\n",
+        "An average insertion time measured by {NUM_INSERTIONS} inserts into an SMT with {size} leaves is {} μs\n",
         // calculate the average
         insertion_times.iter().sum::<u128>() as f64 / (NUM_INSERTIONS as f64),
     );
@@ -100,22 +100,22 @@ pub fn batched_insertion(tree: &mut Smt) -> Result<(), MerkleError> {
 
     let now = Instant::now();
     let mutations = tree.compute_mutations(new_pairs);
-    let compute_elapsed = now.elapsed().as_secs_f64() * 1000_f64;
+    let compute_elapsed = now.elapsed().as_secs_f64() * 1000_f64; // time in ms
 
     let now = Instant::now();
     tree.apply_mutations(mutations)?;
-    let apply_elapsed = now.elapsed().as_secs_f64() * 1000_f64;
+    let apply_elapsed = now.elapsed().as_secs_f64() * 1000_f64; // time in ms
 
     println!(
-        "An average insert-batch computation time measured by a {NUM_INSERTIONS}-batch into an SMT with {size} leaves over {:.1} ms is {:.1} ms",
+        "An average insert-batch computation time measured by a {NUM_INSERTIONS}-batch into an SMT with {size} leaves over {:.1} ms is {} μs",
         compute_elapsed,
-        compute_elapsed / NUM_INSERTIONS as f64,
+        compute_elapsed * 1000_f64 / NUM_INSERTIONS as f64, // time in μs
     );
 
     println!(
-        "An average insert-batch application time measured by a {NUM_INSERTIONS}-batch into an SMT with {size} leaves over {:.1} ms is {:.1} ms",
+        "An average insert-batch application time measured by a {NUM_INSERTIONS}-batch into an SMT with {size} leaves over {:.1} ms is {} μs",
         apply_elapsed,
-        apply_elapsed / NUM_INSERTIONS as f64,
+        apply_elapsed * 1000_f64 / NUM_INSERTIONS as f64, // time in μs
     );
 
     println!(
@@ -163,15 +163,15 @@ pub fn batched_update(tree: &mut Smt, entries: Vec<(RpoDigest, Word)>) -> Result
     let apply_elapsed = now.elapsed().as_secs_f64() * 1000_f64; // time in ms
 
     println!(
-        "An average update-batch computation time measured by a {NUM_UPDATES}-batch into an SMT with {size} leaves over {:.1} ms is {:.1} ms",
+        "An average update-batch computation time measured by a {NUM_UPDATES}-batch into an SMT with {size} leaves over {:.1} ms is {} μs",
         compute_elapsed,
-        compute_elapsed / NUM_UPDATES as f64,
+        compute_elapsed * 1000_f64 / NUM_UPDATES as f64, // time in μs
     );
 
     println!(
-        "An average update-batch application time measured by a {NUM_UPDATES}-batch into an SMT with {size} leaves over {:.1} ms is {:.1} ms",
+        "An average update-batch application time measured by a {NUM_UPDATES}-batch into an SMT with {size} leaves over {:.1} ms is {} μs",
         apply_elapsed,
-        apply_elapsed / NUM_UPDATES as f64,
+        apply_elapsed * 1000_f64 / NUM_UPDATES as f64, // time in μs
     );
 
     println!(
@@ -205,7 +205,7 @@ pub fn proof_generation(tree: &mut Smt) -> Result<(), MerkleError> {
     }
 
     println!(
-        "An average proving time measured by {NUM_PROOFS} value proofs in an SMT with {size} leaves in {:.1} μs",
+        "An average proving time measured by {NUM_PROOFS} value proofs in an SMT with {size} leaves in {} μs",
         // calculate the average
         insertion_times.iter().sum::<u128>() as f64 / (NUM_PROOFS as f64),
     );