… CHANGELOG.md
… Cargo.lock
… Cargo.toml
… src/merkle/smt/full/tests.rs
… src/merkle/smt/mod.rs
… src/merkle/smt/simple/mod.rs