Related to https://github.com/0xPolygonMiden/miden-node/issues/527

While refactoring account inclusion proofs in-memory storage, I realized that it's very hard to keep persistent state for additional "initial" structure, and it would be easier to keep only reverse mutation sets and the latest state. Thus, we need to store only reverse mutations set, but latest state we can still reconstruct from the database.

In this PR I implemented reverse mutations set generation on mutations applying (using a new method `apply_mutations_with_reversion()`. Also implemented serialization of mutations set and added benchmarks for mutations calculation/applying.