@polydez 2024-12-16
> If the effect is significant, I wonder if there is some clean interface to let the user apply mutations with or without generating the reverse set.

One of solutions would be introducing new `feature` for that behavior.