@bobbinth 2024-12-16
> One of solutions would be introducing new `feature` for that behavior.

As discussed offline, having 2 separate methods would be a more flexible solution (because there may be situations where the same codebase may want to use different versions of the method). Maybe, under the hood, both methods would use a single method that could look something like:
```rust
fn apply_mutations(
    &mut self,
    mutations: MutationSet<DEPTH, Self::Key, Self::Value>,
    return_inverse_mutations: bool,
) -> Result<MutationSet<DEPTH, Self::Key, Self::Value>, MerkleError>
```