@polydez 2024-12-17
> Could you run benchmarks to see how this affects things - especially for relatively large trees? (we may already have benchmarks for this).
> 
> If the effect is significant, I wonder if there is some clean interface to let the user apply mutations with or without generating the reverse set. (but let's run the benchmarks first).

I implemented [benchmark](https://github.com/0xPolygonMiden/crypto/pull/355/commits/a5595b9644c7bf3d24d2d7c061cfac00524531a0) for `apply_mutations` and here are the results on MacBook Pro with Apple M3 Max (unnecessary lines were removed):

---

**With** generation of reverse mutations set:

apply_mutations/SimpleSmt: apply_mutations/1000
                        time:   [30.044 ms 30.826 ms 31.697 ms]

apply_mutations/SimpleSmt: apply_mutations/100000
                        time:   [3.7554 s 3.8398 s 3.9335 s]

**Without** generation of reverse mutations set:

apply_mutations/SimpleSmt: apply_mutations/1000
                        time:   [23.827 ms 24.089 ms 24.289 ms]
                        change: [-5.3989% -3.4174% -1.7821%] (p = 0.00 < 0.05)
                        Performance has improved.

apply_mutations/SimpleSmt: apply_mutations/100000
                        time:   [3.2683 s 3.4088 s 3.5618 s]
                        change: [-15.168% -11.224% -6.7039%] (p = 0.00 < 0.05)
                        Performance has improved.

---

Here we can see up to 11% of performance improvement (for 100,000 pairs). The bad news here is that it takes up to 3.6 seconds to apply mutation set (without any other operations, like computations of mutation set). It's higher than current limit per block (65,536 accounts), but I think, it's too slow for fitting in 2 seconds per block limit.