@polydez 2024-12-18
> If this still takes a long time, we may need to start thinking about how to optimize the `apply_mutations()` method.

We should also benchmark `compute_mutations()` method and add metrics for all block building phases because `apply_mutations()` is just a part of block building.