@polydez 2024-12-18
@bobbinth, I have first numbers.

For tree with 1M leaves, benchmark 10K updates:

---
apply_mutations/SimpleSmt: apply_mutations/10000
                        time:   [2.1603 s 2.2503 s 2.3556 s]
---