@bobbinth 2024-12-18
> apply_mutations/SimpleSmt: apply_mutations/10000
> time: [2.1603 s 2.2503 s 2.3556 s]

This is pretty slow (probably around 10x slower than I thought it would be). This basically means we spend 0.1 ms per updated leaf. Could you check what's the bottleneck here?