@polydez 2024-12-19
> > apply_mutations/SimpleSmt: apply_mutations/10000
> > time: [2.1603 s 2.2503 s 2.3556 s]
> 
> This is pretty slow (probably around 10x slower than I thought it would be). This basically means we spend 0.1 ms per updated leaf. Could you check what's the bottleneck here?

The bottleneck here is `BTreeMap` itself. Applying mutations set is just inserting/updating values in `BTreeMap`. There is no duplicates because mutations set contains of two maps with unique keys.

I will try to rewrite `SimpleSmt` to hashmaps, as proposed in one of issues, and measure performance change.