@polydez 2024-12-19
I've rewritten `SimpleSmt` and `Smt` to hashbrown hashmaps, as well, as `MutationSet`. This dramatically improved performance:

---

apply_mutations/SimpleSmt: apply_mutations/10000
                        time:   [154.52 ms 163.09 ms 174.46 ms]
                        change: [-93.242% -92.753% -92.142%] (p = 0.00 < 0.05)
                        Performance has improved.

---

But overall benchmarking takes more than half of hour for 10 samples. So probably `compute_mutations` still takes the most of the time, but this time is not included into the benchmark. I will try to measure it.