@bobbinth 2024-12-19
> I've rewritten `SimpleSmt` and `Smt` to hashbrown hashmaps, as well, as `MutationSet`. This dramatically improved performance:

Very nice! Could you create a separate PR for this? Not sure if possible, but would be awesome to have this behind a feature flag.

Also, we are mostly interested in `Smt` (rather than `SimpleSmt`) as we'll be moving to use `Smt` for both accounts and nullifiers.

> So probably `compute_mutations` still takes the most of the time, but this time is not included into the benchmark. I will try to measure it.

We are working on parallelizing this - so, hopefully, we should be able to run `compute_mutations` on multiple cores.