@bobbinth 2024-12-19
> Do we rely on ordering of iterators over inner nodes or leaves? This is the only obstacle I can see now for full migration of `Smt`/`SimpleSmt` to hashmaps.

Hmmm - good question. I don't think we should be, but I'm not 100% sure we are not.