@bobbinth 2024-12-19
> > Do we rely on ordering of iterators over inner nodes or leaves? This is the only obstacle I can see now for full migration of `Smt`/`SimpleSmt` to hashmaps.
> 
> Hmmm - good question. I don't think we should be, but I'm not 100% sure we are not.

One thing that comes to mind is that order of leaves is important during `Smt` construction (specifically for the multi-threaded version) and so, the order in which we serialize leaves would also be important. So, maybe we do need to preserve the iteration order. But now that I think of it, I'm not sure the current iterators have the same order as what is assumed by the constructor.