@bobbinth 2024-12-19
btw, I think we already have some benchmarks for computing mutations [here](https://github.com/0xPolygonMiden/crypto/blob/next/src/main.rs).