@polydez 2024-12-19
Benchmarking of `compute_mutations()` method (tree size and mutation count are the same, as before: 1M and 10K):

---

compute_mutations/SimpleSmt: compute_mutations/10000
                        time:   [2.6533 s 2.6876 s 2.7379 s]

---


> btw, I think we already have some benchmarks for computing mutations [here](https://github.com/0xPolygonMiden/crypto/blob/next/src/main.rs).

Ah, didn't find this, thank you!